Improve module splitting
